+++ src/main/java/com/linda/trailtracker/dto/UpdateTrailDto.java
@@ -11,7 +11,7 @@ public class UpdateTrailDto {
     private String name;
 
     @Min(value = 1, message = "Distance must be at least 1 km")
-    private int distanceKm;
+    private double distanceKm;
 
     @Min(value = 0, message = "Elevation gain cannot be negative")
     private int elevationGain;
@@ -36,11 +36,12 @@ public class UpdateTrailDto {
         this.name = name;
     }
 
-    public int getDistanceKm() {
+    public double
+    getDistanceKm() {
         return distanceKm;
     }
 
-    public void setDistanceKm(int distanceKm) {
+    public void setDistanceKm(double distanceKm) {
         this.distanceKm = distanceKm;
     }
